# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BrasaFut API - Documentação Oficial",
    "description": "API profissional de futebol de alta performance, com suporte a campeonatos brasileiros e internacionais, dados em tempo real, estatísticas de partidas e webhooks.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "Chave de acesso obtida via /api/v1/auth/register"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/competitions/{id}": {
      "get": {
        "summary": "Obter detalhes de uma competição",
        "tags": [
          "Competições"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "code": {
                      "nullable": true,
                      "type": "string"
                    },
                    "country": {
                      "nullable": true,
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "LEAGUE",
                        "CUP",
                        "INTERNATIONAL"
                      ]
                    },
                    "logoUrl": {
                      "nullable": true,
                      "type": "string"
                    },
                    "createdAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "code",
                    "country",
                    "type",
                    "logoUrl",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/venues/{id}": {
      "get": {
        "summary": "Obter detalhes de um estádio",
        "tags": [
          "Estádios"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/coaches/": {
      "get": {
        "summary": "Listar técnicos de futebol com filtros por clube, nacionalidade e status",
        "tags": [
          "Treinadores & Comissões Técnicas"
        ],
        "description": "Retorna o catálogo completo de treinadores com dados táticos, formação predileta, estatísticas de aproveitamento e clube atual.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "nationality",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "AVAILABLE"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/coaches/ranking": {
      "get": {
        "summary": "Ranking dos melhores técnicos por aproveitamento, títulos ou pontos por jogo",
        "tags": [
          "Treinadores & Comissões Técnicas"
        ],
        "description": "Compara os melhores comandantes do futebol brasileiro e internacional de acordo com seu percentual de vitórias (winRate), quantidade de troféus ou média de pontos (PPM).",
        "parameters": [
          {
            "schema": {
              "default": "winRate",
              "type": "string",
              "enum": [
                "winRate",
                "trophies",
                "pointsPerMatch"
              ]
            },
            "in": "query",
            "name": "sortBy",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/coaches/{id}": {
      "get": {
        "summary": "Perfil completo de um técnico por ID",
        "tags": [
          "Treinadores & Comissões Técnicas"
        ],
        "description": "Apresenta histórico de carreira, DNA tático, cartões amarelos e vermelhos na carreira, aproveitamento em clássicos e títulos conquistados.",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/coaches/{id}/career": {
      "get": {
        "summary": "Linha do tempo da carreira e passagens por clubes de um treinador",
        "tags": [
          "Treinadores & Comissões Técnicas"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/derbies/": {
      "get": {
        "summary": "Listar rivalidades históricas e grandes clássicos",
        "tags": [
          "Grandes Clássicos & Dérbis"
        ],
        "description": "Retorna os maiores clássicos do futebol brasileiro e internacional (Dérbi Paulista, Fla-Flu, Gre-Nal, Clássico Mineiro, El Clásico), incluindo placares históricos e vantagem.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "country",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/derbies/{slug}": {
      "get": {
        "summary": "Estatísticas completas de um clássico específico",
        "tags": [
          "Grandes Clássicos & Dérbis"
        ],
        "description": "Retorna retrospecto geral (vitórias, empates, gols), maiores goleadas de cada lado, maiores artilheiros históricos e últimos 5 confrontos diretos.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/penalties/takers": {
      "get": {
        "summary": "Ranking dos melhores cobradores de pênalti",
        "tags": [
          "Pênaltis & Especialistas"
        ],
        "description": "Lista batedores com maior taxa de conversão (%), aproveitamento sob pressão e zonas prediletas de finalização nas cobranças.",
        "parameters": [
          {
            "schema": {
              "default": 10,
              "type": "number",
              "minimum": 1,
              "maximum": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/penalties/goalkeepers": {
      "get": {
        "summary": "Ranking de goleiros pegadores de pênalti",
        "tags": [
          "Pênaltis & Especialistas"
        ],
        "description": "Goleiros com maior percentual de defesas em penalidades máximas e histórico de decisões por pênaltis vencidas na carreira.",
        "parameters": [
          {
            "schema": {
              "default": 10,
              "type": "number",
              "minimum": 1,
              "maximum": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/penalties/matches/{id}/shootout": {
      "get": {
        "summary": "Auditoria e cobranças de disputa de pênaltis pós-jogo",
        "tags": [
          "Pênaltis & Especialistas"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/attendance/competitions/{id}": {
      "get": {
        "summary": "Ranking de público, bilheteria e ocupação da competição",
        "tags": [
          "Público & Bilheteria dos Estádios"
        ],
        "description": "Relatório financeiro e de engajamento das torcidas: média de público pagante por clube, taxa de ocupação da arena (%), renda bruta acumulada e preço médio do ingresso.",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/attendance/venues/{id}": {
      "get": {
        "summary": "Estatísticas e recorde histórico de público de um estádio",
        "tags": [
          "Público & Bilheteria dos Estádios"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
...
}